+++ e2e/test/scenarios/data-studio/transforms.cy.spec.ts
@@ -4,8 +4,10 @@ import dedent from "ts-dedent";
 
 import { SAMPLE_DB_ID, WRITABLE_DB_ID } from "e2e/support/cypress_data";
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
+import { ORDERS_MODEL_ID } from "e2e/support/cypress_sample_instance_data";
 import type {
   CardType,
+  CollectionItem,
   PythonTransformTableAliases,
   TransformTagId,
 } from "metabase-types/api";
@@ -83,6 +85,43 @@ describe("scenarios > admin > transforms", () => {
       H.assertQueryBuilderRowCount(3);
       H.expectUnstructuredSnowplowEvent({
         event: "transform_created",
+      });
+    });
+
+    it("should not show you the library in the mini picker when building transforms (uxw-2403)", () => {
+      H.createLibrary().then(({ body }) => {
+        cy.request(`/api/collection/${body.id}/items`).then(
+          ({ body: { data: children } }) => {
+            const modelsCollection = children.find(
+              (c: CollectionItem) => c.type === "library-models",
+            ) as CollectionItem;
+
+            cy.request("PUT", `/api/card/${ORDERS_MODEL_ID}`, {
+              collection_id: modelsCollection.id,
+            });
+          },
+        );
+      });
+
+      visitTransformListPage();
+      getTransformsSidebar().button("Create a transform").click();
+      H.popover().findByText("Query builder").click();
+      H.miniPicker().within(() => {
+        cy.findByText(DB_NAME).should("exist");
+        cy.findByText("Our analytics").should("exist");
+        cy.findByText("Browse all").should("exist");
+        cy.findByText("Data").should("not.exist");
+      });
+
+      cy.findByRole("link", { name: "Exit data studio" }).click();
+      H.modal().button("Discard changes").click();
+      H.newButton("Question").click();
+
+      H.miniPicker().within(() => {
+        cy.findByText("Our analytics").should("not.exist");
+        cy.findByText("Browse all").should("exist");
+        cy.findByText("Data").click();
+        cy.findByText("Orders Model").should("exist");
       });
     });
