+++ frontend/src/metabase/common/components/Pickers/MiniPicker/components/MiniPicker/MiniPicker.tsx
@@ -3,18 +3,17 @@ import { useCallback, useEffect, useMemo } from "react";
 import { PLUGIN_DATA_STUDIO } from "metabase/plugins";
 import { Box, Menu } from "metabase/ui";
 
-import { useLogRecentItem } from "../../../EntityPicker/hooks/use-log-recent-item";
-import type { DataPickerValue } from "../../DataPicker";
-import { MiniPickerContext } from "../context";
-import type { MiniPickerItem, MiniPickerPickableItem } from "../types";
+import { useLogRecentItem } from "../../../../EntityPicker/hooks/use-log-recent-item";
+import type { DataPickerValue } from "../../../DataPicker";
+import { MiniPickerContext } from "../../context";
+import type { MiniPickerItem, MiniPickerPickableItem } from "../../types";
 import {
   focusFirstMiniPickerItem,
   getFolderAndHiddenFunctions,
   useGetPathFromValue,
-} from "../utils";
-
-import { MiniPickerListLoader } from "./MiniPickerItemList";
-import { MiniPickerPane } from "./MiniPickerPane";
+} from "../../utils";
+import { MiniPickerListLoader } from "../MiniPickerItemList";
+import { MiniPickerPane } from "../MiniPickerPane";
 
 export type MiniPickerProps = {
   searchQuery?: string;
