+++ frontend/src/metabase/common/components/Pickers/MiniPicker/components/MiniPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./MiniPicker";
